requirements.txt
streamlit>=1.31
plotly>=5.18
pandas>=2.0
sqlalchemy>=2.0
pymysql>=1.1
python-dotenv>=1.0
